path_or_url: Don't load unused remote file modules
Defer the loading and initialization of the remote file support for
individual schemes until it's necessary to satisfy a call to
path_or_url().

Each remote file scheme requires additional Python dependencies (e.g.
HTTPS requires "requests", S3 requires "boto3"), some of which may not
be installed in the workflow's execution environment.

While support for HTTP(S) and S3 is already included in our typical
execution environments, it may not be in all and future schemes will
require additional dependencies.